API

SimDir

ClimaAnalysis.SimDir — Type

SimDir(simulation_path::String)

Object that describes all the ClimaAtmos output found in the given simulation_path.

Base.get — Function

get(simdir::SimDir;
    short_name,
    reduction = "inst",
    period = nothing)

Return a OutputVar for the corresponding combination of short_name, reduction, and period (if it exists).

The variable is read only once and saved into the simdir.

ClimaAnalysis.available_vars — Function

available_vars(simdir::SimDir)

Return the short names of the variables found in the given simdir.

ClimaAnalysis.available_reductions — Function

available_reductions(simdir::SimDir, short_name::String)

Return the reductions available for the given variable in the given simdir.

ClimaAnalysis.available_periods — Function

available_periods(simdir::SimDir, short_name::String, reduction::String)

Return the periods associated to the given variable and reduction.

OutputVar

ClimaAnalysis.OutputVar — Type

Representing an output variable

ClimaAnalysis.read_var — Function

read_var(path::String)

Read a variable in the given NetCDF file.

Example

simdir = SimDir("my_output")
read_var(simdir.variable_paths["hu"]["inst"])

ClimaAnalysis.slice — Function

slice(var::OutputVar, kwargs...)

Return a new OutputVar by slicing across dimensions as defined by the keyword arguments.

Example

slice(var, lat = 30, lon = 20, time = 100)

ClimaAnalysis.slice_general — Function

slice_general(var::OutputVar, val, dim_name)

Return a new OutputVar by selecting the available index closest to the given val for the given dimension

ClimaAnalysis.slice_x — Function

slice_x(var::OutputVar, x)

Return a new OutputVar by selecting the available date closest to the given x.

ClimaAnalysis.slice_y — Function

slice_y(var::OutputVar, y)

Return a new OutputVar by selecting the available date closest to the given y.

ClimaAnalysis.slice_z — Function

slice_z(var::OutputVar, z)

Return a new OutputVar by selecting the available date closest to the given z.

ClimaAnalysis.slice_lon — Function

slice_lon(var::OutputVar, lon)

Return a new OutputVar by selecting the available date closest to the given lon.

ClimaAnalysis.slice_lat — Function

slice_lat(var::OutputVar, lat)

Return a new OutputVar by selecting the available date closest to the given lat.

ClimaAnalysis.slice_time — Function

slice_time(var::OutputVar, time)

Return a new OutputVar by selecting the available snapshot closest to the given time.

ClimaAnalysis.average_lat — Function

average_lat(var::OutputVar)

Return a new OutputVar where the values on the latitudes are averaged arithmetically.

ClimaAnalysis.average_lon — Function

average_lon(var::OutputVar)

Return a new OutputVar where the values on the longitudes are averaged arithmetically.

ClimaAnalysis.average_time — Function

average_time(var::OutputVar)

Return a new OutputVar where the values are averaged arithmetically in time.

Utilities

For development and not

ClimaAnalysis.Utils.match_nc_filename — Function

match_nc_filename(filename::String)

Return short_name, period, reduction extracted from the filename, if matching the expected convention.

The convention is: shortname_(period)_reduction.nc, with period being optional.

Examples

julia> match_nc_filename("bob")

julia> match_nc_filename("ta_1d_average.nc")
("ta", "1d", "average")

julia> match_nc_filename("pfull_6.0min_max.nc")
("pfull", "6.0min", "max")

julia> match_nc_filename("hu_inst.nc")
("hu", nothing, "inst")

ClimaAnalysis.Utils.squeeze — Function

squeeze(A :: AbstractArray; dims)

Return an array that has no dimensions with size 1.

When an iterable dims is passed, only try to squeeze the given dimensions.

Examples

julia> A = [[1 2] [3 4]];

julia> size(A)
(1, 4)

julia> A_squeezed = squeeze(A);

julia> size(A_squeezed)
(4,)

julia> A_not_squeezed = squeeze(A; dims = (2, ));

julia> size(A_not_squeezed)
(1, 4)

ClimaAnalysis.Utils.nearest_index — Function

nearest_index(A::AbstractArray, val)

Return the index in A closest to the given val.

Examples

julia> A = [-1, 0, 1, 2, 3, 4, 5];

julia> nearest_index(A, 3)
5

julia> nearest_index(A, 0.1)
2

ClimaAnalysis.Utils.kwargs — Function

kwargs(; kwargs...)

Convert keyword arguments in a dictionary that maps Symbols to values.

Useful to pass keyword arguments to different constructors in a function.

Examples

julia> kwargs(a = 1)
pairs(::NamedTuple) with 1 entry:
  :a => 1

CairoMakie

ClimaAnalysis.Visualize.heatmap2D! — Function

heatmap2D!(fig::CairoMakie.Figure,
           var::ClimaAnalysis.OutputVar;
           p_loc = (1,1),
           more_kwargs)

Plot a heatmap of the given 2D variable in the given figure and location.

The plot comes with labels, units, and a colorbar.

This function assumes that the following attributes are available:

long_name

short_name

units (also for the dimensions)

Additional arguments to the plotting and axis functions

more_kwargs can be a dictionary that maps symbols to additional options for:

the axis (:axis)

the plotting function (:plot)

the colorbar (:cb)

The values are splatted in the relevant functions. Populate them with a Dictionary of Symbols => values to pass additional options.

ClimaAnalysis.Visualize.sliced_heatmap! — Function

sliced_heatmap!(
                    fig::CairoMakie.Figure,
                    var::ClimaAnalysis.OutputVar,
                    cut::Union{Nothing, AbstractDict{String, <: Real}};
                    p_loc = (1,1),
                    more_kwargs,
                    )

Take a variable, slice as directed, and plot a 2D heatmap in the given figure and location.

The plot comes with labels, units, and a colorbar.

Arguments

If the variable is not 2D, cut has to be a dictionary that maps the dimension that has to be sliced and the value where to cut.

For example, if var has four dimensions: time, long, lat, z, this function can be used to plot a lat-long heatmap at fixed time and z. Assuming we want to plot time 100. and altitude 50., cut should be Dict("time" => 100., "z" => 50.).

This function assumes that the following attributes are available:

long_name

short_name

units (also for the dimensions)

Additional arguments to the plotting and axis functions

more_kwargs can be a dictionary that maps symbols to additional options for:

the axis (:axis)

the plotting function (:plot)

the colorbar (:cb)

The values are splatted in the relevant functions. Populate them with a Dictionary of Symbols => values to pass additional options.

ClimaAnalysis.Visualize.heatmap! — Function

heatmap!(
            fig::CairoMakie.Figure,
            var::ClimaAnalysis.OutputVar;
            p_loc = (1,1),
            more_kwargs,
            kwargs...
            )

Syntactic sugar for sliced_heatmap with kwargs instead of cut.

Example

heatmap!(fig, var, time = 100, lat = 70) plots a heatmap by slicing var along the time nearest to 100 and latitude nearest 70.

Additional arguments to the plotting and axis functions

more_kwargs can be a dictionary that maps symbols to additional options for:

the axis (:axis)

the plotting function (:plot)

the colorbar (:cb)

The values are splatted in the relevant functions. Populate them with a Dictionary of Symbols => values to pass additional options.

ClimaAnalysis.Visualize.line_plot1D! — Function

line_plot1D!(
             fig::CairoMakie.Figure,
             var::ClimaAnalysis.OutputVar;
             p_loc = (1,1),
             more_kwargs
             )

Plot a heatmap of the given 2D variable in the given figure and location.

The plot comes with labels, units, and a colorbar.

This function assumes that the following attributes are available:

long_name

short_name

units (also for the dimensions)

Additional arguments to the plotting and axis functions

more_kwargs can be a dictionary that maps symbols to additional options for:

the axis (:axis)

the plotting function (:plot)

The values are splatted in the relevant functions. Populate them with a Dictionary of Symbols => values to pass additional options.

ClimaAnalysis.Visualize.sliced_line_plot! — Function

sliced_line_plot!(
                    fig::CairoMakie.Figure,
                    var::ClimaAnalysis.OutputVar,
                    cut::Union{Nothing, AbstractDict{String, <: Real}};
                    p_loc = (1,1),
                    more_kwargs
                    )

Take a variable, slice as directed, and plot a 1D line plot in the given figure and location.

The plot comes with labels, and units.

Arguments

If the variable is not 1D, cut has to be a dictionary that maps the dimension that has to be sliced and the value where to cut.

For example, if var has four dimensions: time, long, lat, z, this function can be used to plot a lat-long heatmap at fixed time and z. Assuming we want to plot time 100. and altitude 50., cut should be Dict("time" => 100., "z" => 50.).

This function assumes that the following attributes are available:

long_name

short_name

units (also for the dimensions)

Additional arguments to the plotting and axis functions

more_kwargs can be a dictionary that maps symbols to additional options for:

the axis (:axis)

the plotting function (:plot)

The values are splatted in the relevant functions. Populate them with a Dictionary of Symbols => values to pass additional options.

ClimaAnalysis.Visualize.sliced_plot! — Function

sliced_plot!(
             fig::CairoMakie.Figure,
             var::ClimaAnalysis.OutputVar,
             cut::Union{Nothing, AbstractDict{String, <: Real}};
             p_loc = (1,1),
             more_kwargs
            )

Take a variable, slice as directed, and plot a 1D line plot or 2D heatmap in the given figure and location.

The plot comes with labels, and units (and possibly a colorbar).

Arguments

If the variable is not 1D/2D, cut has to be a dictionary that maps the dimension that has to be sliced and the value where to cut.

For example, if var has four dimensions: time, long, lat, z, this function can be used to plot a lat-long heatmap at fixed time and z. Assuming we want to plot time 100. and altitude 50., cut should be Dict("time" => 100., "z" => 50.).

This function assumes that the following attributes are available:

long_name

short_name

units (also for the dimensions)

Additional arguments to the plotting and axis functions

more_kwargs can be a dictionary that maps symbols to additional options for:

the axis (:axis)

the plotting function (:plot)

the colorbar (:cb)

The values are splatted in the relevant functions. Populate them with a Dictionary of Symbols => values to pass additional options.

ClimaAnalysis.Visualize.plot! — Function

plot!(
      fig::CairoMakie.Figure,
      var::ClimaAnalysis.OutputVar;
      p_loc = (1,1),
      more_kwargs,
      kwargs...
      )

Syntactic sugar for sliced_plot with kwargs instead of cut.

Example

line_plot!(fig, var, time = 100, lat = 70) plots a line plot or a heatmap by slicing var along the time nearest to 100 and latitude nearest 70.

Additional arguments to the plotting and axis functions

more_kwargs can be a dictionary that maps symbols to additional options for:

the axis (:axis)

the plotting function (:plot)

the colorbar (:cb)

The values are splatted in the relevant functions. Populate them with a Dictionary of Symbols => values to pass additional options.